app/requirements.txt
uWSGI==2.0.19.1
blinker==1.5
certifi==2022.9.24
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
sentry-sdk==1.9.9
urllib3==1.26.12
Werkzeug==2.2.2
python-logstash==0.4.8